\name{ITEMAN}
\alias{ITEMAN}
\alias{ITEMAN}
\docType{package}
\title{
ITEMAN: Classical Item Analysis
}
\description{
This package
}
\details{
\tabular{ll}{
Package: \tab ITEMAN\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-09-29\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
The package can be used to run classical item analysis for multiple-choicee test items and polytomously scored items (e.g., rating scale items).
}
\author{
Maintainer: Cengiz Zopluoglu <c.zopluoglu@miami.edu>
}
\keyword{ package }
\seealso{
\code{\link{ITEMAN1}} for running classical item analysis for multiple-choice test items
\code{\link{ITEMAN2}} for running classical item analysis for polytomously scored items
}